src/main/java/leetcode/string/MultiplyStrings.java
package leetcode.string;

//Given two non-negative integers num1 and num2 represented as strings, return the product of num1 and num2, also represented as a string.
//
//        Example 1:
//
//        Input: num1 = "2", num2 = "3"
//        Output: "6"
//        Example 2:
//
//        Input: num1 = "123", num2 = "456"
//        Output: "56088"
//        Note:
//
//        The length of both num1 and num2 is < 110.
//        Both num1 and num2 contain only digits 0-9.
//        Both num1 and num2 do not contain any leading zero, except the number 0 itself.
//        You must not use any built-in BigInteger library or convert the inputs to integer directly.

import java.util.Objects;

public class MultiplyStrings {
    public String multiply(String num1, String num2) {
        if (num1 == null || num2 == null) {
            throw new NullPointerException("Inputs cannot be null");
        }
        if (num1.isBlank() || num2.isBlank()) {
            throw new IllegalArgumentException("Inputs cannot be blank");
        }

        if (num1.equals("0") || num2.equals("0")) {
            return "0";
        }

        int len1 = num1.length();
        int len2 = num2.length();
        int[] res = new int[len1 + len2];
        for (int i = len1 - 1; i >= 0; i--) {
            int n1 = num1.charAt(i) - '0';
            for (int j = len2 - 1; j >= 0; j--) {
                int n2 = num2.charAt(j) - '0';

                int sum = (res[i + j + 1] + n1 * n2);
                res[i + j] += sum / 10;
                res[i + j + 1] = sum % 10;
            }
        }

        StringBuilder result = new StringBuilder();
        for (int i = 0; i < res.length; i++) {
            if (i == 0 && res[i] == 0) continue;
            result.append(res[i]);
        }

        return result.toString();

    }
}
